moved npy.h into test directory.
