+++ include/demo.h
@@ -10,7 +10,6 @@
 #include <vector>
 
 #include "draw/draw.h"
-#include "npy.h"
 #include "pcdet-infer-cpu/common/metadata.h"
 #include "pcdet-infer-cpu/common/runtimeconfig.h"
 #include "pcdet-infer-cpu/pcdet.h"
