+++ tests/snapshot_shape_test.cpp
@@ -3,8 +3,8 @@
 
 #include <numeric>
 
-#include "npy.h"
 #include "pcdet_test/include/config.h"
+#include "pcdet_test/include/npy.h"
 #include "pcdet_test/include/params.h"
 #include "pcdet_test/include/pre.h"
 #include "pcl.h"
